refactor(lib): move subtitle related code to subtitle folder
