+++ src/index.ts
@@ -1,8 +1,7 @@
 export * from './components';
-export * from './cue';
 export * from './drm';
 export * from './events';
 export * from './hooks';
 export * from './player';
 export * from './source';
-export * from './subtitleTrack';
+export * from './subtitle';
